requirements.txt
Django==3.0.7
django-environ==0.4.5
psycopg2==2.8.5
django-model-utils==4.0.0
djangorestframework==3.11.0
django-extensions==2.2.9

isort
black

pytest==5.4.3
pytest-django==3.9.0
freezegun==0.3.15